pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "f1-2026-engine"
version = "0.1.0"
description = "Production-grade Formula 1 2026 hybrid strategy simulation engine."
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "black>=24.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.black]
target-version = ["py311"]
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
